Renamed 'regs' to 'labels'.
I think that Greg was right to call regs labels. This patch renames the
corresponding methods and method call in all parts of the source that already
survived the transition.

Please note, that this also affects former Dataset.reglabels and
Dataset.chunklabels, which are now called Dataset.unique{labels,chunks}