Added interfaces (and examples) of SplitProcessor() and ErrorFx().
I made a further Base class for ErrorFx() called ErrorFxBase(). The latter
makes no assumptions about the arguments of the __call__() method.
ErrorFx() instead takes the predicted and the desired values of a
classification/regression as arguments to its __call__() method.

I think that in most cases this is the desired interface for an error function.

Moreover I switched the name of the SplitProcessing class to SplitProcessor.
I like it better this way ;-)