Drafted new CrossValidation class.
Old crossval code has moved into mvpa/oldcrossval.py.

Removed the 'predictions' from the argument list of the SplitProcessor __call__() method. It limits the CrossValidation class as it would have to know where in the return tuple of the Splitter the validation set is stored (if any second dataset is available at all).