Neighbor classes. More TODO
Might need to be refactored. Finders actually have nothing to do with
Mappers, including MaskMapper. MaskMapper (or derived class) can
implement Finder interface, and it will be done via delegation to some
instance of the finder to get points around a given feature. then mask
have to be applied to remove Features which do not belong to the mask.