Modifed class hierarchy for Mappers.
Removed MaskMetricMapper and replaced it by MetricMapper which in turn is the
baseclass for MaskMapper. I believe that a mask array has always some kind of
inherent metric (even if only the element indexes). Therefore the MaskMapper
should implement the Metric interface. I went a step further and made the
DescreteMetric with cartesianDistance the default Metric for MaskMapper.

This modification makes the implicite metrics of a mask array (as implied by
element indexing) also available as an explicite metric available to any
MaskDataset.

No significant code modification are caused by this modification as MaskMapper
and MaskDataset continue to work as usual.