NF: Say hello to the first implementation of NiftiDataset.
It is based on MaskedDataset and set its mapper automatically while making
use of the NIfTI voxdim property to determine the appropriate metric.

It stores the full NIfTI object inside so it should be easy to provide mapper
wrapper that transform some results directly into NIfTI file with the same
properties.