Alternative implementation of a loop in maskmapper.__init__.
It should be the same as the loop and actually should be faster
although consecutive check if a voxel belongs to the mask has to be in
getOutId function call. Since profiling showed that this location
seems not to be a battleneck, this fix for now resides in a branch.