RF+BF: Refactored SVM interface -> more specific subclasses. Tests bugfix.
The goal behind this refactoring is to make SVM classifiers more intuitive and
easier to maintain. The old code requires users to know about libsvm behavior.
The new code is an attempt to completely encapsulate the SVM functionality and
hide libsvm internals from the user while keeping the flexibility.

Additonally a proper SVM clf hierarchy was necessary. ATM we only have the
possibility to compute SVM weights for linear SVMs. Therefore we need a
SensitivityAnalyzer for linear SVMs and a different one for non-linear SVMs.
Having a common LinearSVM baseclass makes things more transparent IMHO.

There was also a little bugfix in the SVM unittests (error in dataset).