RF+NF: Refactored IO helpers to be read & write; Class for SampelAttributes IO
The IO helpers for ColumnData are no targeted to be really input _and_ output
helpers. FSL's EV3 format now can be read and written.

Initial draft of a class to handle IO of Dataset sample attributes. Don't know
how complex it should become. Currently it simply reads two columns, where the
fist is assumed to be the label and the second is the chunk of a sample.

Adapted and enhanced unit tests.

Also the searchlight example is adjusted to use the fresh SampleAttributes
helper. ATTENTION: Please resync example dataset, because I removed the header
from the attributes file. We need to agree on a proper format. Maybe register
our own MIME type ... ;-)