RF: Initial draft of 'TransferError'.
This is a first sketch of the formerly known ErrorOracle. At the time of this
writing there is still need for more thought. I don't know yet whether this is a
bright idea or if it will go away tomorrow.

Perhaps it is a good idea to _not_ merge right now. Anyway, all unit tests are
adjusted accordingly and pass.