BF: Fixed TailFeatureSelector to do what its documentation suggests ;-)
TailFeatureSelector.__call__() docstring says: 'Call function returns Ids
to be kept'. BUt it actually removed sensitivity values from the passed
sensitivity map.

I guess behaviour was intended for some reason and I feel like I destroyed
something. But RFE was coded in a way that it expects Ids to be kept and that
is what TailFeatureSelector does now.

Unit tests are adjusted accordingly.

RFE itself seems to work now. Previously reported bug is fixed with this
commit.