RF: FeatureSelector -> ElementSelector
Let the selectors work on upper and lower tails both to discard and to select
elements. Not sure yet, whether this is a good implemention. However, it is
needed for a simple implmentation of 'IncrementalFeatureSearch', which should
be as similar as to RFE as possible.

Unit test and RFE are adjusted accordingly.