NF: Added some simple wrapper to transform the output of our many callables.
First two implementations are needed to take the absolute of sensitivity map
and flip the output of the searchlight algorithm to be used as a sensitivity
analyser.

These 'transformers' are intended to be used as wrappers that can be directly
given to the contructors of the relevant algorithm that needs transformed input.

Most likely all of these transformations can be done by simple lambda functions,
but I like them being a concrete member of the code ;-)

Corresponding unit tests are included.