NF: Website demo from rST + CSS.
Played a bit with rST and found it really nice to generate a webpage. The
included page might become the future PyMVPA website.

I prefer to keep it very simple: only one main page plus some links to other
generated sources, e.g. API docs.

There is a new makefile target 'website' that calles rst2html to generate the
webpage from source. The generated HTML is XHTML-compliant -- nice!

Added python-docutils to the build-depends of the Debian package as we might
want to included the webpage in the shipped docs as HTML.

Really not to mention that style is debatable...  ;-)