RF: Combiner, inherit instead to delegate in MulticlassClassifier and SplitClassifier
Previously it was delegation since Yarik thought to create a separate
BoostedClassifier which would combine not via simple combiner such as
MaximalVote but through another classifier. Now Yarik adjusted
Combiner API so we can easily create a Combiner which trains itself on
the results of training of each particular classifier. And then uses
that to come up with the solution. It is TODO but now should be
possible.

Thus now MulticlassClassifier and SplitClassifier are true kids of
CombinedClassifier, which simplified a few things