RF: now specific wrappers/adapters for corresponding classifiers are under clfs.<library>
sg is still the default but commented out line is placed for making
libsvm default

XXX think if such wrappers to external packages shouldn't be placed
under some common mvpa.ext. For instance later on we might make use of
MDP and that one provides not only classifiers so looking for it under
clfs might be confusing. Or may be we could spread things around?