Finished one pass through manual.
I made a number of minor rewrites and ran spellcheck, but I'm not sure if some of the words that were
corrected were due to Brittish vs. American English.  I left all the neighbour spellings in there because
I think it looks cool :)

Anyway, there's obviously still a lot missing so I'll certainly help out more on it.