NF: SensitivityAnalyzer example + fMRI example dataset
Example runs same dataset on a bunch of SensitivityAnalyzers. It is done in the
spirit of pylab_2d, this time showing sensitivity maps.

Example dataset is very minimal (to keep size down). Only posterior half of a
ventral, axial slice (not detrended, but motion corrected, single subject).
Origin is Haxby (2001) 8-category dataset. This is the same experiment that
serves as example for the MVPA toolbox (but different subject).

It may serve as dataset for examples, but might also help for unit tests.

Adjusted .gitignore to not ignore the dataset.