BF: Possibly fixed bug in NiftiDataset
I think there is a larger issue with a possibly flawed coding practice used throughout pymvpa.
Specifically, there loads of times where a default value of a keyword argument is set to be either {} or
[].  I think the proper way to deal with this is to have the value set to None and then manually set it
in the code to the empty dict or list.  Otherwise, the default values for a method are only evaluated
once, which means that you are reusing the reference to the empty dict/list, which I think was causing
all the issues with NiftiDataset and could be causing other memory issues, as well.

So, as of now, all the tests pass when I changed out the dsattr to default to None and not {}.  I think
we should really verify whether it's ok to use {} and [] as default kwarg values and if not, these should
be changed throughout the entire package.