NF: Started devguide section about 'Extending PyMVPA'
IMHO we are really lacking such documentation. It is scattered over in many
source files and already starts to diverge from the reality. I'd like to
introduce this devguide section as the authorative definition of our external
and internal interfaces. It should help us to maintain stable versions of
them and also increase transparency of decisions of the past. (I have to
confess that I cannot recall every bit and piece of the changes that were
introduced -- even by myself ;-)