NF/RF: Added new SMLR zero-weight checking method plus weights a property
Worried that we were not revisiting zero weights often enough with the previous
implementation, I changed the way the weights get polled.  It now is configurable based on a
decay to a minimum revisitation probability.  This is slightly slower than what we had
before, but it may give rise to more consistent results across multiple calls on the same
dataset (please let me know how this turns out for you.)

As for the property vs. state debate, Michael and I decided to make weights and biases
properties because they will always be there.  See the devguide for more discussions on
this.  I have fixed the smlrweights and all the tests pass.