NF: Harvester -- grab anything from any number of anything that ran on anything ;-)
Took Yarik's draft of Loop, killed the beauty and added some more stuff to it.
Now it is a complicated monster, but we would probably not primarily give it to
the end-user. However, it should be really powerful when used inside of some
convenience classes. I added a few close to real-world unit tests.