NF/BF: Minor bug fixed in SMLR and added ability to fit all weights.
So, when we want to actually make sense out of the weights in a multiclass problem, we don't want SMLR to
only fit weights for the M-1 classes (although it is equivalent).  So, I added an option to
fit_all_weights, which then fits a weight to each feature for each class.  This means that the weights
that are left at the end are actually meaningful :)

To do: fix the smlrweights to not take the mean over classes because we probably want to take the mean
over the abs of the classes to get the most useful weights.  I'll talk to Yarik and Michael about this.