BF: include copy of copy.py from python2.5 for pymvpa used with python<2.5
Otherwise deepcopy fails to copy a class which has a binding to a
function (in our case - combiners for sensitivity analyzers). This
issue seems to be fixed with copy.py from python2.5.1-7. If used
python is less than 2.5, this copy is used instead of system-wide
installed one.

whenever python 2.4 is deprecated enough we can avoid shipping copy.py