NF: Harvestable base class
Introduced to overcome problem of absent access to internals of
classifiers from some high level abstractions such as
CrossValidatedTransferError, as well as a helper within
CombinedClassifier. So if it is needed to collect some information at
some point which would be discarded in the next loop iteration. Thus
it is imho cleaner (or at least easier to use) alternative to
callbacks whenever it is desired to accumulate some information