RF+BK: Evil commit: selectAnalyzer is gone...
I made an attempt to remove selectAnalyzer which cause me headaches (and
circular dependencies, eg. 'import smlr' caused all SVM implementations
to be imported as well). This situation should be resolved now....

Classifier got a new methods 'getSensitivityAnalyzer(**kwargs) which has
to be implemented for each classifier to return the matching
SensitivityAnalyzer. It seems to work -- although this commit only implements
it for SMLR.

Additonally the unit test battery and some additional pieces are modified to
run even without *any* SVM implementation.