NF/BF/BK/RF: Added new dependency test code.
This new code allows for testing individual dependencies without testing for all
dependencies.  For example, there's no reason to start a session with R via rpy if you
are not going to be using any of the R-based classifiers.

I replaced all the code and all the tests that were passing now pass.  Once you have
OK'd this change we can remove the comments that I left in there with the old method of
dependency checking.