NF: Proposal for enriched doc strings for our classes.
This should make it easier to comprehend what a class is doing and what it
needs as constructor arguments. Especially for derived classes, this helper
removes the need to search for base class docs. It also interacts quite nicely
with epydoc.

If you want to try, do this in ipython:

	>>> from mvpa.measures.searchlight import Searchlight
	>>> Searchlight?

and also

	make apidoc
	www-browser build/html/api/mvpa.measures.searchlight.Searchlight-class.html

To enable this feature for your class you need to put this into the file:

	from mvpa.base.dochelpers import enhancedDocString

and in the class (right _after_ the constructor):

	__doc__ = enhancedDocString('NameOfTheClass', locals(), BaseClass)

BaseClass is optional.

This seems to be the only way to do this, as one cannot modify the doc string
of a class after its creation.