NF+RF+BK: Start transition to 'Sphinx' for all documentation handling.
Sphinx is a pretty amazing system form docs in RsT format. It can build our
website with HTML docs and also builds pretty PDFs. It is not totally bug-free,
but quite useable and provides a number of advantages compared to our previous
approach:

  * Index generation
  * Offline search capabilities
  * Easy cross-references across documents
  * Same tool that generates the new python docs -> people are familiar with it

This commit starts to move our docs under the control of sphinx. The transition
is not yet done, but you can already get a feeling for the new look. In the
toplevel dir do:

	make htmldoc
	make pdfdoc

and

	kpdf build/latex/*.pdf
	konqueror build/html/index.html

Hope you like it.