NF+RF: Proposal to integrate assessment of statistical significance of transfer errors into `TransferError`
It is done by having an optional class instance that estimates the NULL
distribution of transfer error (currently only brute-force monte carlo
implemented). `TransferError` will automatically estimate the NULL dist
whenever such instance is given and also a training dataset is available.

If a NULL dist is estimated transfer error will query the cdf of that
distribution to determine the probability of a certain transfer error under
the NULL dist and will store the result in `null_prob` state variable.

TODO: think about integrating that information in CrossValidatedTransferError.