RF+BK?: Evil, intrusive change to Mapper class hierarchy (please review)
Previouse mapper hierarchy looked like this:

Mapper -> MetricMapper -> MaskMapper
       |
	   -> SVDMapper

I think this design was broken, because it did not allow to build
mappers that need to do selection of elements (which _all_ mappers
should do) without inheriting the metrics. Currently, I want to create a
Mapper from a space with no inherent metrics. I realized that what I
want is MaskMapper (A mapper that selects data elements based on a
binary mask) without metrics.

Therefore I think we should have:

Mapper -> MaskMapper -> MetricMapper -> DenseArrayMapper

where last one behaves like MaskMapper used to do.

I think this is reasonable, as we will need 'element-selection based on
a binary mask' for a number of other mappers as well (with or without
metrics), e.g. SparseArrayMapper, MeshMapper, ....

However, even if I'm right, this is a very intrusive change. E.g. for
some reason (I cannot recall right now), the old MaskMapper implemented
__deepcopy__ and always returned 'MaskMapper' instances. Now that it
moved in the class hierarchy this does not work anymore. For now I
commented it out, but I'm sure that it caused problems at some point.
Where do we need to copy the Mappers? The unittests run fine -- so
whatever has been the problem in the past is not tested ATM!

Please review my changes. I did not do them in a separate branch as they
are somewhat central to my plan for the near future and I would prefer
things fixed in the mainline.

BTW: Of course, without __deepcopy__ things run twice as fast.