BF: Make SMLR extension functional under win32.
The problemi was the following: SMLR is compiled by distutils as an
extension, but in a technical sense it is none -- it is just a shared
lib that is loaded with ctypes. However, under windows extensions
become .pyd thingies.

I 'fixed' it be making smlr a dummy extension and exported the
'stepwise_regression' symbol manually and added an explicite loader for this
function under win32.