NF: Registry-monster for PYMVPA.
This is an attempt to merge all ways to change the runtime behavior of PyMVPA
into a single entity (and their documentation as well).

Advantages are that we provide the user with the possibility to configure
user- and analysis-specifc settings using simple textfiles, while keeping
are current approach using env variables.

The configuration is exposed via `mvpa.cfg` and can be queried/modified by
library and user-code. Another nice feature is, that users can now easily dump
the configuration they used for an analysis into a file by doing::

	mvpa.cfg.save('filename.cfg')

and re-use/inspect it later on.

Additonally, we expose a logic to handle any configuration options and
therefore save users from having to do it themselves.

Unittest fragment is added, but very useful ATM.

So far mvpa.cfg is only used to query and set the RNG seed, because you might
not like having a registry ;-)