BF: reset to proper basic prior sigma_p
When BLR was instanced and trained on a dataset of a given dimension
it crashed when re-used on another dataset having different
dimensions.  This happened because sigma_p did not match the correct
new dimension anymore.  A simple check on dimension-value has been
added that reset sigma_p to an identity matrix of the correct
dimensions in case of mismatches with current 'data'.

This solves the errors in standard tests, but remember that not
setting properly sigma_p can lead to wrong results!