NF: Persistent samples ids: 'origids' samples attribute; few comments.
Each dataset now generates unique sampels ids, ie. 'origids' which
survive calls to selectSamples(). This will put us into the position to
compute sample-wise stats (e.g. misclassification rate to identify the
'problematic' ones). Till now, this information was lost as we are using
first-axis-position in the samples array to identify each sample.

An alternative to 'origids' would be using a mapper to get 'coordinates'
in InSpace, but this would only work for MappedDataset and makes it more
difficult to develop generic code for sample-wise stats, ie. InSpace
coordinates might be anything (at least in theory).