NF: added gradients for KernelConstant and KernelLinear + cleanup + renaming
Added compuation of gradient of kernel and log marginal likelihood in
KernelConstant and KernelLinear (method: compute_lml_gradient()).

Removed some useless functions and added new ones in Kernel to raise
NotImplementedError when gradients are not available for some kernels.

Renaming of a variable k_grad_i -> K_grad_i which is a more meaningful
name.