NF+DOC : addedd 'kernel_width' as hyperparameter + improved docs
As stated in YiJun Sun 2007 PAMI paper the kernel computation assume
an extra "kernel_width" hyperparameter, since I-RELIEF weights sum
up to 1. According to the author this hyperparameter is not critical
to the result.

Added extra documentation to remember that current supported kernels
are only the exponential-like ones, and not all, currently. Added
other minor docs about complexity etc.