DOC+NF: First draft of a docstring converter.
The converter takes our docstring (which are supposed to be pretty human
readable) and enhances them with sphinx markup to beautify the generated
API docs.

This is a very minimal version that handles parameter definition lists
and return value description. But it can easily be extended to enhance
XXX comments or even strip them from the docs ;-)