BF: AUC now averaged across splits with nanmean for confusion matrix.
I changed this from mean because occasionally you may get a split that has only one type of
label in the test set, which would give rise to a nan.  This single nan shouldn't keep you
from getting an average AUC in your confusion matrix.

I added the nanmean function to mvpa.clfs.stats for when scipy is not installed.