BF: provide cvtranserror ability to be used with RFE-based classifier
also BF: unbind testdataset within SplitClassifier

as an example in clfs.warehouse, it was recommended to use
SplitClassifier around classifier which uses RFE for feature
selection. CrossValidatedTransferError wasn't able to handle such
cases (ie to provide classifier with testdataset for possible testing)
and is introducing generalization estimation bias -- since stopping
point would be determined based on the testing part of the dataset.

This BF provides such functionality within CrossValidatedTransferError
in a fashion similar to SplitClassifier. That again shows the need to
refactor two beast for better code reuse.