RF: Avoid DeprecationWarnings with latest pynifti.
Unfortunately there is no way to test for 0.20081017.1 but trying to
import the nifticlib module -- which is in turn the one that throws the
Warning in 0.20090205.1.

I decided to simply sacrifice 0.20081017.1 in favor of 0.20090205.1 --
doesn't hurt anyway. Of course a more sophisticated test function will
have done it is well -- lazy me.