NF/RF/BF: Regularization now parameter of GPR.
I'm not sure what to call it, so I put it all.  Anyway, I just added the regularization
parameter to be used all the time by GPR.  It still defaults to 0.0, so there should be no
change for people currently using GPR with no issues.  However, this enables the manual
setting of the regularization if you run into the not positive, definite error.  I figured
that if you were going to use regularization, you would want to use it all the time and not
have different values on different folds of your data.