BF: test_svm -- recrreate cv since clf might not be bound anymore
this is an echo of the same issue -- whenever transerrors state is
enabled, cvte operates on copies.

Most sensible solution would be to adjust cvte so for the last split it
operates on original of transerror, but for that splitters have to expose
either configurations they currently yield on, or number of those.  But it
somewhat brakes the notion that splitters do not expose those 'states'