NF: Draft of a new FeatureSubsetMapper.
This is the second half of the tandem FlattenMapper+FeatureSubsetMapper
that aims to replace the old MaskMapper.

The most important improvement is (aside from being a lot simpler) that
is no longer has issues with the order of feature ids passed to
select_out() and also natively support all kinds of Numpy slicing
styles.