RF: SimpleCollection (currently duplicate of Collection)
SimpleCollection is an extract of Collection that is necessary to
provide the functionality required for Dataset. It is much simpler
than Collection itself (which remains as a base class for
ParameterCollection and States). Ideally, Collection would be refactored
to use SimpleCollection as its base class ... time will bring more
clarity.

I feel that SimpleCollection should: 1. get a better name and 2. should
really feel like a dictionary.