NF/OPT: check if filtering really needed in Sphere
Previously we would attempt to filter all coordinates in the Sphere __call__.
Now we do a simple check to see if the largest and smallest coordinates are
outside the range, and only filter if this is the case. This should increase
performance since most of the time the sphere is inside the volume, and so
we only ever need to filter coordinates at the edges and corners. Some simple
checks with %timeit in ipython confirm the speed gain.