NF: Parallelized Searchlight.
Simplistic approach to parallelize the searchlight. It uses the pprocess
module to utilize a set of processes to compute the sphere measure.

The whole thing does not scale to well. On my dual-core machine I get
approx. 25% speed-up for 2 processes, in comparison to just one (using
the searchlight_2d example as a benchmark.

The same benchmark in a cluster node with 8 cores yields 250% speedup
when utilizing all 8 cores.