RF+BK: Merge combiners and transformers into a mapper for DatasetMeasures.
The whole design was flawed anyway (just look at the comments in
mvpa/measures/base). Now the measure should return a plain 1D array, or
a Dataset. The latter can easily deal with base-sensitivties and the
like. The DatasetMeasure base-class can finally apply a _single_ mapper
to any measure. The might even be a chain mapper. Any mapper can deal
with 1D-arrays (forward1()) and datasets (forward()).

We might need to implement more mapper the handle some of the
combiner/transformer combinations that we need -- but at least sticking
to the mapper concept will help to reduce the complexity here.