BF/RF:  normalizer_cls/args + tolerate older shogun version to some degree
Added normalizer arguments to SGKernel and none is assigned by
default. Also for PolySGKernel -- make it work with older versions of
shogun which specified normalization flag to the constructor.

Unfortunately there seems to be no sane way to provide
PrecomputedSGKernel for those older versions, so as_ng conversions
would puke