RF+BK: Trying to update PrototypeMapper, but failed.
It used the old Kernel interface. I changed that only to discover that
Similarity also use (or rather mimic) the old interface. I changed that
too, to discover that there are difficulties to feed arbitrary objects
into a Dataset. I eventually got that to work, but finally failed to get
StreamlineSimilarities computed.

Emanuele, could you take a look?

Once this is done the test needs to be included in the test battery to
prevent thins like this from happening!