BF: repliminary fix to avoid test to fail.
When prototypes are a list and samples are an ndarray of objects the
samples extraction does not work well and wrong data arraves to the
similarity computation function. This fix creates prototypes as an
ndarray of objects which solve the issue.

To Michael & Yaroslav: which is the more generic form of Dataset we
accept? just iterable containers or something better like ndarray of
objects? After some thoughts I guess we should choose ndarray of
objects, nothing less.