RF+NF: ZScoreMapper to replace its ancestor and `zscore`. Close gh-23
The new mapper offers the same Mapper+function combo as
PolyDetrendMapper + poly_detrend(). It can do what the old zscore()
function could do, but now in a proper Mapper-way. This API is
different, and can deal with sample attributes in a more flexible way.

Right now the `inspace` argument is unused.

Since this is a rather critical piece of infrastructure, a code review
would be most welcome.