RF+BIG: Rename 'labels' -> 'targets'. Close gh-16
This should much better reflect the purpose of this attribute for e.g.
regressions.

The commit also renames related stuff:

uniquelabels -> uniquetargets
           L -> T
          UL -> UT

This is a huge diff. There is every chance that I missed something or
fucked up some hidden piece. Anyway, the full unittests pass (with
shogun installed, but without RPy{1,2}), as well as the doctests.

Let's hope we find the remaining pieces.