RM: CA feature_ids
Used to store the feature ids that were used to train a classifier.
However, this wasn't really reliable anymore once any mapper can modify
the feature set on any meta classifier level. Instead classifier should
get the ability to store this information, based on actual feature ids of
an incoming dataset, in an outgoing result dataset as a dataset
attribute (or whatever...)