Temporarily change Collection.is_enabled()
Not is also checks whether an attribute is known at all and returns
False if not -- instead of IndexError raise before. This is done to
keep many 'is_enabled('feature_ids')' from puking, because there is no
common CA like this anymore. Shall be removed once the whole issue is
completely refactored.