NF(RF): explicit argument 'flatten=None' to from_wizard -- default behavior change
before flattening was unconditionally done even if the mapper provided, which
was assumed to opearate only after masking+flattening.  But in some cases it is
handy to use wizard to create a dataset having already precrafted (or borrowed)
mapper which does everything including flattening.  So, from now on by default
if mapper is provided, no flattening would be done if no mask is given, and
original dataset would be passed  to the mapper.