BF: introduce rcdf() to address the issue of incorrectly low p-value for right-tail testing
especially prominent if null-chance distribution is degenerate
(e.g. the same value) and the actual estimate is of the same
value. Then used before 1-cdf(x) is incorrect way since we need to add
pmf(x) into this equation for discrete samples (for continuous
variables/distributions it remains ok to take just 1-cdf)