NF: Add scatter_neighborhood()
Neighborhood seeds (or centers) are placed on coordinates drawn from a
provided list so that no seed is part of any other neighborhood. Depending
on the actual shape and size of the neighborhoods, their elements can be
overlapping, only the seeds (or centers) are guaranteed to be
non-overlapping with any other neighborhood. This can be used to perform
sparse sampling of a given space.