BF: memory leak in libsvm bindings
swig does not take into account LIBSVM_VERSION #define because it is
supposed to come from svm.h which it does not parse.  So that prevents
svm_destroy_model_helper being created.

And before models were not destroyed at all leading to heavy memory
leaking.  Tested with one with included libsvm 2.89 (?) and system-wide
installed libsvm 3.0 on Michael"s test script -- according to valgrind
leaks seems to be gone and unittests passed on squeeze