PY3: Makefile to run tests out of py3k build directory if provided PYTHON=python3
That makes it possible to avoid custom testing/build rules for python3
(although existing ones are left in for now).  And it should be
sufficient simply to provide necessary python version in PYTHON
variable.  The only catch (as it now done customly in a rule for
unittest-py3) is to remove build stamp if sources changed since
testing done is out of source for python3, e.g.:

  rm -f build-stamp ; make unittest-badexternals PYTHON=python3

As a "side-effect" it would even leave everything working as destined
if default python on the system is of version 3 ;)