RF: Improve scoping of hyperalignment code
Each level is now handle in a dedicated method. Garbage collection
automatically removed the left-overs from previous levels. This should
reduce the memory footprint of some steps (e.g. list of mapped datasets
is not needed in level3). It also improve the redability of the code
as all levels, their inputs and outputs are no clearly identified.