BF: adjusting ttest_1samp code and test to pass on older numpy/scipy's
apparently numpy/scipy bundle on debian squeeze is problematic (see
comments in the code for this commit).  So tests had to be relaxed a
bit to not create the case which would lead to a failure (e.g. having
multi-dimensional array with only 2 elements not masked-out is some
row/column).