ENH: SL -- results_fx and nblocks to allow "on-the-fly" processing of results
with nblocks >> nproc and custom results_fx it becomes feasible to
work on problems where storing all searchlight results in memory and
post-processing them all at once is prohibitive due to large memory
requirements.

with this modification input is split into larger number of smaller
feature "partitions", results spit out by a generator which is passed
to results_fx making it possible to process them "on-the-fly" and
dumping them right after thus not requiring as much memory.