NF: Wrap sklearn's dataset generators/loaders
Adds a bunch of skl_*() to the suite that share the exact same
interface as their sklearn.datasets parents, but return a PyMVPA
dataset.

Unittest verifies that the wrapper signatures are identical, and that
some minimum amount of functions were wrapped -- which will hopefully
detect when sklearn changes its naming convention.