NF(+RF): AttributePermutator got option strategy with new handling for uattrs value
if strategy="uattrs" is specified, then attributes are not simply
permuted breaking any possible existing interdependency -- samples
of the same unique combination of attrs obtain the same attrs value
of another unique combination.  In the simplest case, e.g. if
attrs=targets, it results in simply remapping targets (i.e. e.g.
1 becomes 2, 2 becomes 1 and 0 stays 0).

The idea is that this should be closer to the idea of "permutting
independent samples" as "clean" permutation MC H0 testing requires and
hopefully would eliminate the necessity for the scary two-stage design
to not permute "testing" labels.  With this setup, in case of balanced
labels, it should be ok to simply reassign labels in all functional runs
gaining similarly concervative H0 distribution (yet to be tested)