ENH: when unable to resolve ties -- choose first one in sorted order
on one hand it is just making it somewhat worse (e.g. before it could be
considered random resolution since order of labels comes from a unordered dictionary)
but this way we should obtain results consistent with LIBSVMs failure to resolve
ties.  So -- for now lets be more consistent with LIBSVM.  Correct resolution still TODO