ENH: expose Multiclass's ca.raw_predictions also as a ca.raw_predictions_ds with correct orientation
Felt wrong to break meaning of what/how things are presented in raw_predictions.
But classifier x sample was not a good fit for pass_attr, so came up with
new .ca.  Having it as a Dataset allows to disambiguate what targets
actually corresponded to any given binary classifier's 'positive/negative' labels