ENH: first shout at RFELearner -- helper to run most common RFE analysis
based on splitting training data and running nested CV with RFE
to figure out optimal number of features to stay with, and then
running RFE on full training dataset.

TODO: most probably separate "Learner" is not necessary and we should just get
a derived FeatureSelection to be used with a FeatureSelectionClassifier to avoid
too much of custom constracts (e.g. for sensitivities)