BF+ENH: FxMapper to make output order of groups deterministic (sorted by default)
Also allows for order=occurence which should sort groups in "original" order
as samples within the groups appeared in the dataset. Would be valuable
to see run-order dependencies after block averaging etc.  Ideally that should
even be default but that would change behaviour of code of many, where now
order is determined by order of keys in a dictionary, thus not fully
deterministic to say the least.  And that is what caused daily tests breakage
with Python 3