ENH: Include the 1slice example dataset in openfmri.org layout
This is just the start. The 25mm version can be added as an additional
flavor. The reason for doing this is to have a dataset that isn't
concatenated runs, but has them separate as most real-world datasets do.
Second, this is to promote the idea of having data organized according
to the openfmri scheme, including a specification of the experiment
design.