BF: indexing bug with searchlight_hyperalignment
Fixing a bug that could cause `ValueError` when using `ref_ds` along with `exclude_from_model`.
Fixing a minor bug with `ref_ds` when using the same QueryEngine for all datasets.
Now the `ref_ds` parameter of `SearchlightHyperalignment` is the index before excluding any datasets, and would override the `ref_ds` parameter of `Hyperalignment`.
An explicit test is added for `exclude_from_model`.
The documentation has also been updated and is consistent with the code.