BF: set affine to the one we know in map2nifti
Original comment hinted that it was supposed to be set by nibabel if provided
affine is None.  I have tested in precise docker image with nibabel 1.1.0 (2011ish)
and it (already) was not the case.  So better to set it explicitly and test for
that in the tests.  Thanks @ddwagner for the report and analysis!

Closes #599