BF: GNB - add has_sensitivity tag in constructor
has_sensitivity (per our initial discussion) was decided to be assigned
at train time since .params.common_variance could be changed.  But that
precluded that instance to be tested as a part of the warehouse for
classifiers with sensitivities. So here I change it to assign in
constructor and just assert in the actual call that it is all still
consistent.  That now enables testing GNB with the generic test of
sensitivities (previous commit also established thorough testing for
GNB sensitivities)