BF: guard against under/overflows and mitigate in case of logprob=True (default)
we mitigate by offseting all the values by the maximal value (in log space),
computing marginals, and then offsetting back when back in the logspace.

Similar strategy should be possible for operating in original, non logspace.
But may be there is even a better way, so not bothering for now.

Closes #581